caching/python/requirements.txt
Flask
Flask-SQLAlchemy
redis
psycopg2-binary
python-dotenv